Refactor from /app router to /pages router
Current build uses a hybrid. /app router handles /api routing, while /pages router handles the page rendering.
Most of it is the same, except that I removed /src route and changed Tailwind config to match.
Updated all imports to use @/ instead of ../../ to target root level directories.
Renamed /typings to /types
Removed "use client" on all page routes (might still be on components, harmless).
Removed all /src references.

Tested captcha and it's working as intended. Since in the front-end we are still targeting the same route /api/verify-captcha on the front-end, the functionality is the same.