Fix aspect ratio scaling issue (not clickbait)
Aspect ratio logic is removed from front-end as we are now we are capturing the native camera
The aspect ratio handling has been moved to the backend where we are scaling the videos to 9:16.

Fixes weird crop-in effect on desktop and tablet, as well as tablet black bars issue.

This code currently uses a test environment I've setup on AWS, so next step is to rewrite this to use the correct bucket and copy settings over to prod environment.